Move text conditioning into pocket_tts/modules/text_conditioner.py
The conditioners package held a 7-line base.py (TokenizedText) and a single
text.py; a one-file module under modules/ matches how the rest of the
package is laid out.
