--- pocket_tts/conditioners/base.py
@@ -1,7 +1,0 @@
-from typing import NamedTuple
-
-import torch
-
-
-class TokenizedText(NamedTuple):
-    tokens: torch.Tensor  # should be long tensor.
