+++ training/modules/conditioner.py
@@ -9,7 +9,7 @@ import torch
 import torch.nn.functional as F
 from torch import nn
 
-from pocket_tts.conditioners.base import TokenizedText
+from pocket_tts.modules.text_conditioner import TokenizedText
 
 from ..args import TrainArgs
